+++ py2app_tests/basic_app/__init__.py
@@ -1,0 +1,1 @@
+"""dummy init file to trick 2to3"""